+++ ovp_projects/views.py
@@ -1,4 +1,5 @@
 from django.core.exceptions import ObjectDoesNotExist
+from django.db.models import Q
 
 from ovp_projects.serializers import project as serializers
 from ovp_projects import models
@@ -45,6 +46,13 @@ class ProjectResourceViewSet(mixins.CreateModelMixin, mixins.RetrieveModelMixin,
       instance = self.get_object()
       serializer = self.get_serializer(instance)
 
+    return response.Response(serializer.data)
+
+  @decorators.list_route(['GET'])
+  def manageable(self, request, *args, **kwargs):
+    projects = models.Project.objects.filter(Q(owner=request.user) | Q(organization__owner=request.user) | Q(organization__members=request.user))
+
+    serializer = self.get_serializer_class()(projects, many=True)
     return response.Response(serializer.data)
 
   @decorators.detail_route(['POST'])
@@ -118,6 +126,9 @@ class ProjectResourceViewSet(mixins.CreateModelMixin, mixins.RetrieveModelMixin,
     if self.action == 'retrieve':
       self.permission_classes = ()
 
+    if self.action == 'manageable':
+      self.permission_classes = (permissions.IsAuthenticated, )
+
     return super(ProjectResourceViewSet, self).get_permissions()
 
   def get_serializer_class(self):
@@ -127,5 +138,7 @@ class ProjectResourceViewSet(mixins.CreateModelMixin, mixins.RetrieveModelMixin,
       return serializers.ApplyCreateSerializer
     if self.action == 'applies':
       return serializers.ApplyRetrieveSerializer
+    if self.action == 'manageable':
+      return serializers.ProjectRetrieveSerializer
 
     return serializers.ProjectRetrieveSerializer